// codes/frontend/src/types/servercommands.ts
export enum ServerCommand {
	NEW_PLAYER = "NEW_PLAYER",
	BET_TIME_START = "BET_TIME_START",
	INSURANCE_TIME_START = "INSURANCE_TIME_START",
	PLAYER_TIME_START = "PLAYER_TIME_START",
	LOGIN = "LOGIN",
	ROUND_COMPLETED = "ROUND_COMPLETED",
	PLAYER_LEFT = "PLAYER_LEFT",
	STATUS_UPDATE = "STATUS_UPDATE",
	OPEN_TABLE = "OPEN_TABLE",
	TIMED_OUT = "TIMED_OUT",
	INITIAL_DEAL_DONE = "INITIAL_DEAL_DONE",
	NO_BETS_NO_DEAL = "NO_BETS_NO_DEAL",
	SHOWDOWN = "SHOWDOWN",
	INIT_DATA = "INIT_DATA",
	SPINNING = "SPINNING"
}